+++ app/models/tests.server.ts
@@ -90,14 +90,3 @@ export async function deleteTestById(id: string) {
     },
   })
 }
-
-export async function getQuestionById(id: string) {
-  return prisma.question.findUnique({
-    where: { id },
-    include: {
-      options: true,
-      correctAnswer: true,
-      correctOptions: true,
-    },
-  })
-}
